Test of Repeated Values

This function analyzes the frequency with which values get repeated within a set of numbers. Unlike Benford's law, and its generalizations, this approach examines the entire number at once, not only the first or last digit.

rv.test(x, alternative = 'two.sided', check = 'last', method = 'af', B = 1000)

Arguments

x
a numeric vector of values from which the digits should be analyzed.

alternative
a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".

check
which digits to shuffle during the procedure. Can be last or lasttwo.

method
which property of the data is calculated. Defaults to af for average frequency, but can also be entropy for entropy.

B
how many samples to use in the bootstraping procedure.

Value

An object of class dt.rv containing:

x

input data.

frequencies

frequencies of observations in x.

samples

vector of simulated samples.

integers

counts for extracted integers.

decimals

counts for extracted decimals.

n

the number of observations in x.

statistic

the value the average frequency or entropy statistic.

p.value

the p-value for the test.

cor.test

correlation test for the integer portions of the number versus the decimals portions of the number.

method

method used.

check

checked digits.

data.name

a character string giving the name(s) of the data.

Details

To determine whether the data show an excessive amount of bunching, the null hypothesis that the data do not contain an unexpected (random) amount of repeated values is tested. If alternative = "greater" the alternative is that x has more repeated values than expected, and if alternative = "less" the alternative is that x has less repeated values than expected.The statistic can either be the average frequency (\(AF = sum(f_i^2)/sum(f_i))\) of the data or the entropy (\(E = - sum(p_i * log(p_i))\), with \(p_i=f_i/n\)) of the data. Average frequency and entropy are highly correlated, but the average frequency is often more interpretable. For example, an average frequency of 2.5 means that, on average, your observations contain a value that appears 2.5 times in the data set.To quantify what is expected, this test requires the assumption that the integer portions of the numbers are not associated with their decimal portions.

References

Simohnsohn, U. (2019, May 25). Number-Bunching: A New Tool for Forensic Data Analysis. Retrieved from http://datacolada.org/77.

See also

Author

Koen Derks, k.derks@nyenrode.nl

Examples

 

x <- rnorm(100)

# Repeated values analysis shuffling last digit
rv.test(x, check = 'last', method = 'af')

#> 
#> 	Repeated values test
#> 
#> data:  x
#> n = 100, AF = 1, p-value = 1
#> alternative hypothesis: frequencies of repeated values are different than for random data.

# Repeated values analysis shuffling last two digits
rv.test(x, check = 'lasttwo', method = 'entropy')

#> 
#> 	Repeated values test
#> 
#> data:  x
#> n = 100, S = 4.6052, p-value = 2
#> alternative hypothesis: frequencies of repeated values are different than for random data.